+++ internal/circle/service.go
@@ -4,10 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/google/uuid"
 	"github.com/stellar/go/strkey"
-
-	"github.com/stellar/stellar-disbursement-platform-backend/internal/services/assets"
 
 	"github.com/stellar/stellar-disbursement-platform-backend/internal/utils"
 )
@@ -19,6 +16,8 @@ type Service struct {
 	EncryptionPassphrase string
 }
 
+const StellarChainCode = "XLM"
+
 // ServiceInterface defines the interface for Circle related SDP operations.
 //
 //go:generate mockery --name=ServiceInterface --case=underscore --structname=MockService --output=. --filename=service_mock.go --inpackage
@@ -27,50 +26,7 @@ type ServiceInterface interface {
 	SendPayment(ctx context.Context, paymentRequest PaymentRequest) (*Transfer, error)
 }
 
-type PaymentRequest struct {
-	SourceWalletID            string
-	DestinationStellarAddress string
-	Amount                    string
-	StellarAssetCode          string
-	IdempotencyKey            string
-}
-
-func (p PaymentRequest) getCircleAssetCode() (string, error) {
-	switch p.StellarAssetCode {
-	case assets.USDCAssetCode:
-		return "USD", nil
-	case assets.EURCAssetCode:
-		return "EUR", nil
-	default:
-		return "", fmt.Errorf("unsupported asset code: %s", p.StellarAssetCode)
-	}
-}
-
-func (p PaymentRequest) Validate() error {
-	if p.SourceWalletID == "" {
-		return fmt.Errorf("source wallet ID is required")
-	}
-
-	if !strkey.IsValidEd25519PublicKey(p.DestinationStellarAddress) {
-		return fmt.Errorf("destination stellar address is not a valid public key")
-	}
-
-	if err := utils.ValidateAmount(p.Amount); err != nil {
-		return fmt.Errorf("amount is not valid: %w", err)
-	}
-
-	if p.StellarAssetCode == "" {
-		return fmt.Errorf("stellar asset code is required")
-	}
-
-	if err := uuid.Validate(p.IdempotencyKey); err != nil {
-		return fmt.Errorf("idempotency key is not valid: %w", err)
-	}
-
-	return nil
-}
-
-var _ ServiceInterface = &Service{}
+var _ ServiceInterface = (*Service)(nil)
 
 type ServiceOptions struct {
 	ClientFactory        ClientFactory
@@ -136,7 +92,7 @@ func (s *Service) SendPayment(ctx context.Context, paymentRequest PaymentRequest
 		},
 		Destination: TransferAccount{
 			Type:    TransferAccountTypeBlockchain,
-			Chain:   "XLM",
+			Chain:   StellarChainCode,
 			Address: paymentRequest.DestinationStellarAddress,
 		},
 	})
